adjustment in data type

+++ lib/core/domain/entities/rating_model.dart
@@ -9,7 +9,7 @@ class RatingModel{
 
   factory RatingModel.fromJson(Map<String, dynamic> json){
     return RatingModel(
-      rate: json['rate'],
+      rate: json['rate'].toDouble(),
       count: json['count'],
     );
   }
